pyproject.toml
[project]
name = "credo-webhooks-client"
version = "0.1.0"
description = "Lightweight Python client for managing webhooks with the Credo AI API"
authors = [
    {name = "Nick Cemenenkoff"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = "^3.10"
dependencies = [
    "requests>=2.31.0,<3.0.0",
    "python-dotenv>=1.0.0,<2.0.0"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
